+++ src/main/java/xt9/deepmoblearning/common/trials/affix/ITrialAffix.java
@@ -13,5 +13,4 @@ public interface ITrialAffix {
     void run();
     String getAffixName();
     String getAffixNameWithFormatting();
-    String getAffixDescription();
 }